requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
smartapi-python==1.4.2
pyotp==2.9.0
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
httpx==0.27.0
python-dotenv==1.0.1
apscheduler==3.10.4
loguru==0.7.2
pydantic==2.7.1
pytz==2024.1
requests==2.31.0
